pyproject.toml
[project]
name = "chorus-hub"
version = "0.1.0"
description = "Discord-to-Claude Code session router — Hub component"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3",
    "aiohttp>=3.9",
    "click>=8.1",
    "nest-asyncio>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-aiohttp>=1.0",
    "ruff>=0.4",
]

[project.scripts]
chorus = "hub.main:cli"

[project.urls]
Homepage = "https://github.com/zxiang77/chorus"
Issues = "https://github.com/zxiang77/chorus/issues"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["hub"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
